+++ src/app/tools/gen-mapper/dialogs/location-dialog/location-dialog.component.ts
@@ -1,10 +1,27 @@
+import { MapsAPILoader } from '@agm/core/services/maps-api-loader/maps-api-loader';
 import { Component, ElementRef, Inject, NgZone, ViewChild } from '@angular/core';
 import { AbstractControl, FormControl } from '@angular/forms';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
-import { MapsAPILoader } from '@agm/core/services/maps-api-loader/maps-api-loader';
+import { MapsService } from '@core/maps.service';
 
 interface MouseEvent {
     coords: { lat: number, lng: number };
+}
+
+export interface LocationDialogConfig {
+    placeId?: string;
+    address?: string;
+    latitude?: number;
+    longitude?: number;
+    markerLatitude?: number;
+    markerLongitude?: number;
+}
+
+export interface LocationDialogResponse {
+    address?: string;
+    placeId?: string;
+    latitude?: number;
+    longitude?: number;
 }
 
 @Component({
@@ -23,25 +40,37 @@ export class LocationDialogComponent {
     public searchControl: AbstractControl;
     public zoom: number;
     public address: string;
+    public placeId: string;
     public height: number;
     public geocoder: google.maps.Geocoder;
 
     constructor(
+        private mapsService: MapsService,
         private mapsAPILoader: MapsAPILoader,
         private ngZone: NgZone,
         private dialogRef: MatDialogRef<LocationDialogComponent>,
-        @Inject(MAT_DIALOG_DATA) public data: { coords: Coordinates }
+        @Inject(MAT_DIALOG_DATA) public data: LocationDialogConfig
     ) {
-        this.latitude = data.coords.latitude;
-        this.longitude = data.coords.longitude;
+        this.placeId = data.placeId;
+        this.address = data.address;
+        this.latitude = data.latitude;
+        this.longitude = data.longitude;
+        this.markerLatitude = data.markerLatitude;
+        this.markerLongitude = data.markerLongitude;
+
         this.zoom = 12;
-        this.searchControl = new FormControl();
+        this.searchControl = new FormControl(this.address);
         this.height = window.innerHeight - 150;
         this.initialize();
     }
 
     public onSubmit(): void {
-        this.dialogRef.close(this.address);
+        this.dialogRef.close({
+            address: this.address,
+            placeId: this.placeId,
+            latitude: this.markerLatitude,
+            longitude: this.markerLongitude
+        });
     }
 
     public markerDragEnd(event: MouseEvent): void {
@@ -55,9 +84,7 @@ export class LocationDialogComponent {
     private initialize(): void {
         this.mapsAPILoader.load().then(() => {
 
-            const autocomplete = new google.maps.places.Autocomplete(this.searchElementRef.nativeElement, {
-                types: ['address']
-            });
+            const autocomplete = new google.maps.places.Autocomplete(this.searchElementRef.nativeElement, {});
 
             autocomplete.addListener('place_changed', () => {
                 this.ngZone.run(() => {
@@ -66,7 +93,7 @@ export class LocationDialogComponent {
                     if (place.geometry === undefined || place.geometry === null) {
                         return;
                     }
-
+                    this.placeId = place.place_id;
                     this.address = place.formatted_address;
                     this.latitude = place.geometry.location.lat();
                     this.longitude = place.geometry.location.lng();
@@ -76,7 +103,19 @@ export class LocationDialogComponent {
                 });
             });
 
-            this.setAddress(this.latitude, this.longitude);
+            if (this.address && !this.markerLatitude) {
+                this.mapsService.getCoordsForAddress({ address: this.address, placeId: this.placeId }).subscribe(result => {
+                    this.ngZone.run(() => {
+                        this.placeId = result.placeId;
+                        this.latitude = result.latitude;
+                        this.longitude = result.longitude;
+                        this.markerLatitude = this.latitude;
+                        this.markerLongitude = this.longitude;
+                    });
+                });
+            } else if (!this.markerLatitude) {
+                this.setAddress(this.latitude, this.longitude);
+            }
         });
     }
 
@@ -88,6 +127,7 @@ export class LocationDialogComponent {
         this.geocoder.geocode({ 'latLng': latLng } as google.maps.GeocoderRequest, (res, status) => {
             if (status === google.maps.GeocoderStatus.OK) {
                 if (res[0]) {
+                    this.placeId = res[0].place_id;
                     this.address = res[0].formatted_address;
                     this.searchControl.setValue(res[0].formatted_address);
                 } else {