+++ src/app/tools/gen-mapper/gen-mapper/gen-mapper.component.ts
@@ -5,13 +5,14 @@ import { AuthenticationService } from '@core/authentication.service';
 import { Unsubscribable } from '@core/Unsubscribable';
 import { DocumentDto } from '@shared/entity/document.model';
 import { FileInputDialogComponent } from '@shared/file-input-dialog/file-input-dialog.component';
+import { cloneDeep, some } from 'lodash';
 import { takeUntil } from 'rxjs/operators';
 
 import { CreateDocumentDialogComponent } from '../dialogs/create-document-dialog/create-document-dialog.component';
+import { GenMapperGraphComponent } from '../gen-mapper-graph/gen-mapper-graph.component';
+import { GenMapperViewTabsComponent } from '../gen-mapper-view-tabs/gen-mapper-view-tabs.component';
 import { GMTemplate, GNode, PrintType } from '../gen-mapper.interface';
 import { GenMapperService } from '../gen-mapper.service';
-import { GenMapperGraphComponent } from '../gen-mapper-graph/gen-mapper-graph.component';
-import { cloneDeep } from 'lodash';
 import { NodeClipboardService } from '../node-clipboard.service';
 
 @Component({
@@ -23,11 +24,15 @@ export class GenMapperComponent extends Unsubscribable implements OnInit {
     @ViewChild(GenMapperGraphComponent)
     public genMapperGraph: GenMapperGraphComponent;
 
+    @ViewChild(GenMapperViewTabsComponent)
+    public tabs: GenMapperViewTabsComponent;
+
     public template: GMTemplate;
     public node: GNode;
     public document: DocumentDto;
     public documents: DocumentDto[];
     public isAuthenticated: boolean;
+    public showMapView: boolean;
 
     constructor(
         private authService: AuthenticationService,
@@ -51,6 +56,10 @@ export class GenMapperComponent extends Unsubscribable implements OnInit {
                 if (!result.document && !this.authService.isAuthenticated() && this.genMapper.hasLocalDocument()) {
                     this.router.navigate([this.template.name, 'local'], { skipLocationChange: true });
                 }
+
+                if (this.document) {
+                    this.showMapView = some(this.document.nodes, d => !!d.location);
+                }
             });
 
         this.genMapper.getNode()
@@ -69,7 +78,13 @@ export class GenMapperComponent extends Unsubscribable implements OnInit {
     public onGraphChange(nodes: GNode[]): void {
         this.document.nodes = nodes;
         this.genMapper.updateDocument(this.document)
-            .subscribe(result => { console.log('updated'); });
+            .subscribe(result => {
+                console.log('updated');
+                this.document = cloneDeep(this.document);
+                // Changing the document reference triggers a change.
+                this.genMapper.setDocument(this.document);
+                this.showMapView = some(this.document.nodes, d => !!d.location);
+            });
     }
 
     public onNodeClick(node: GNode): void {
@@ -77,7 +92,16 @@ export class GenMapperComponent extends Unsubscribable implements OnInit {
     }
 
     public onUpdateNode(node: GNode): void {
-        this.genMapperGraph.graph.updateNode(node);
+        const nodeToUpdate = this.document.nodes.find((d: any) => d.id === node.id);
+        if (nodeToUpdate) {
+            Object.assign(nodeToUpdate, node);
+
+            if (this.genMapperGraph) {
+                this.genMapperGraph.graph.redrawData(this.document.nodes);
+            } else {
+                this.onGraphChange(this.document.nodes);
+            }
+        }
     }
 
     public onCopyNode(node: GNode): void {