+++ src/app/tools/gen-mapper/node-drawer/node-drawer.component.ts
@@ -1,18 +1,19 @@
-import { Component, OnInit, Input, OnChanges, Output, EventEmitter } from '@angular/core';
-import { GNode, GMField, GMTemplate } from '../gen-mapper.interface';
-import { GenMapperService, GenMapperConfig } from '../gen-mapper.service';
-import { Unsubscribable } from '@core/Unsubscribable';
-import { takeUntil, take } from 'rxjs/operators';
-import { TemplateUtils } from '../template-utils';
+import { Component, EventEmitter, Input, OnChanges, OnInit, Output } from '@angular/core';
+import { FormControl, FormGroup } from '@angular/forms';
+import { MatDialog, MatDrawer } from '@angular/material';
 import { LocaleService } from '@core/locale.service';
+import { Unsubscribable } from '@core/Unsubscribable';
 import { DocumentDto } from '@shared/entity/document.model';
-import { MatDrawer, MatDialog } from '@angular/material';
-import { cloneDeep, without, assign } from 'lodash';
-import { FormGroup, FormControl } from '@angular/forms';
 import { FileInputDialogComponent } from '@shared/file-input-dialog/file-input-dialog.component';
+import { assign, cloneDeep } from 'lodash';
+import { takeUntil } from 'rxjs/operators';
+
 import { ConfirmDialogComponent } from '../dialogs/confirm-dialog/confirm-dialog.component';
+import { GMField, GMTemplate, GNode } from '../gen-mapper.interface';
+import { GenMapperService } from '../gen-mapper.service';
 import { NodeClipboardService } from '../node-clipboard.service';
 import { PeopleGroupService } from '../people-group.service';
+import { TemplateUtils } from '../template-utils';
 
 @Component({
     selector: 'app-node-drawer',
@@ -33,6 +34,9 @@ export class NodeDrawerComponent extends Unsubscribable implements OnInit, OnCha
 
     @Input()
     public template: GMTemplate;
+
+    @Input()
+    public hideActions: boolean;
 
     @Output()
     public pasteNode = new EventEmitter<GNode>();
@@ -137,6 +141,7 @@ export class NodeDrawerComponent extends Unsubscribable implements OnInit, OnCha
 
         this.updateNode.emit(this.node);
         this.drawer.disableClose = false;
+        this.form.reset();
         this.genMapper.setNode(null);
     }
 
@@ -163,7 +168,7 @@ export class NodeDrawerComponent extends Unsubscribable implements OnInit, OnCha
 
     public onImportSubtree(): void {
         this.dialog
-            .open(FileInputDialogComponent, { minWidth: '400px' })
+            .open(FileInputDialogComponent, { minWidth: '350px' })
             .afterClosed()
             .subscribe(result => {
                 if (result) {
@@ -199,6 +204,12 @@ export class NodeDrawerComponent extends Unsubscribable implements OnInit, OnCha
 
         // Add custom control for parentId
         group.parentId = new FormControl(null);
+
+        if (group.location) {
+            group.placeId = new FormControl(null);
+            group.latitude = new FormControl(null);
+            group.longitude = new FormControl(null);
+        }
 
         this.form = new FormGroup(group);