+++ src/app/tools/gen-mapper/templates/church-circles.ts
@@ -163,6 +163,21 @@ export const ChurchCirclesTemplate = {
             'type': 'geoLocation'
         },
         {
+            'header': 'latitude',
+            'initial': null,
+            'type': 'hidden'
+        },
+        {
+            'header': 'longitude',
+            'initial': null,
+            'type': 'hidden'
+        },
+        {
+            'header': 'placeId',
+            'initial': '',
+            'type': 'hidden'
+        },
+        {
             'header': 'date',
             'initialTranslationCode': 'initialDate',
             'type': 'text',