+++ src/main.ts
@@ -11,12 +11,23 @@ import { translations } from './app/tools/gen-mapper/templates/translations';
 import { environment } from './environments/environment';
 import { Device } from '@core/platform';
 import { Browser } from '@core/browser';
+import { ChurchCirclesCzechTemplate } from './app/tools/gen-mapper/templates/church-circles-czech';
+import { DisciplesTemplate } from './app/tools/gen-mapper/templates/disciples';
 
 if (environment.production) {
     enableProdMode();
 }
 
-const resources = defaultsDeep(translations, defaultsDeep(ChurchCirclesTemplate.translations, FourFieldsTemplate.translations));
+let resources = translations;
+
+[
+    ChurchCirclesTemplate.translations,
+    FourFieldsTemplate.translations,
+    ChurchCirclesCzechTemplate.translations,
+    DisciplesTemplate.translations
+].forEach(translation => {
+    resources = defaultsDeep(resources, translation);
+});
 
 i18next
     .use(i18nextBrowserLanguageDetector)
@@ -26,7 +37,12 @@ i18next
     });
 
 platformBrowserDynamic().bootstrapModule(AppModule)
-    .then(() => {
+    .then((ref) => {
+        // Ensure Angular destroys itself on hot reloads.
+        if (window['ngRef']) {
+            window['ngRef'].destroy();
+        }
+        window['ngRef'] = ref;
         Device.setClassList(document.documentElement);
         Browser.setClassList(document.documentElement);
     })